refactor(fbuild-python): split lib.rs into submodules to satisfy LOC gate
Splits the 2094-line crates/fbuild-python/src/lib.rs into eight topic-focused
modules so every .rs file in this crate is under 900 LOC, satisfying the new
1000-LOC gate. No public API change: the #[pymodule] entry point, all
#[pyclass] types, and all #[pyfunction] registrations are preserved exactly.
