+++ crates/fbuild-python/src/json_rpc.rs
@@ -1,0 +1,161 @@
+//! REMOTE: JSON-RPC response helpers and shared async read/write loops
+//! used by both `SerialMonitor` and `AsyncSerialMonitor`.
+
+use base64::Engine;
+use futures::{SinkExt, StreamExt};
+use std::sync::Arc;
+use tokio_tungstenite::tungstenite;
+
+use crate::messages::{ClientMessage, ServerMessage, WsSink, WsSource};
+
+pub(crate) fn extract_remote_json_rpc_response(lines: &[String]) -> Option<String> {
+    lines.iter().find_map(|line| {
+        line.strip_prefix("REMOTE:")
+            .map(|json_part| json_part.to_string())
+    })
+}
+
+pub(crate) fn wait_for_remote_json_rpc_response<F>(timeout: f64, mut poll: F) -> Option<String>
+where
+    F: FnMut(f64) -> Vec<String>,
+{
+    let deadline = std::time::Instant::now() + std::time::Duration::from_secs_f64(timeout);
+
+    while std::time::Instant::now() < deadline {
+        let remaining = (deadline - std::time::Instant::now()).as_secs_f64();
+        let lines = poll(remaining);
+        if let Some(json_part) = extract_remote_json_rpc_response(&lines) {
+            return Some(json_part);
+        }
+    }
+
+    None
+}
+
+/// Shared async read-batch loop used by `AsyncSerialMonitor::read_lines`
+/// and `write_json_rpc`. Acquires the source mutex only for the duration
+/// of each `.next()` call so that concurrent `write` / `__aexit__`
+/// futures can still progress between iterations.
+pub(crate) async fn read_lines_async(
+    ws_read_slot: Arc<tokio::sync::Mutex<Option<WsSource>>>,
+    auto_reconnect: bool,
+    timeout_secs: f64,
+) -> Vec<String> {
+    let mut lines = Vec::new();
+    let deadline = std::time::Instant::now() + std::time::Duration::from_secs_f64(timeout_secs);
+
+    loop {
+        let now = std::time::Instant::now();
+        if now >= deadline {
+            break;
+        }
+        let remaining = deadline - now;
+
+        let mut guard = ws_read_slot.lock().await;
+        let Some(source) = guard.as_mut() else {
+            // Session not entered (or already exited). Nothing to read.
+            break;
+        };
+
+        let result = tokio::time::timeout(remaining, source.next()).await;
+        // Drop the guard before continuing the loop so another task
+        // can take the mutex (e.g. __aexit__).
+        drop(guard);
+
+        match result {
+            Ok(Some(Ok(tungstenite::Message::Text(text)))) => {
+                match serde_json::from_str::<ServerMessage>(&text) {
+                    Ok(ServerMessage::Data {
+                        lines: data_lines, ..
+                    }) => {
+                        lines.extend(data_lines);
+                        if !lines.is_empty() {
+                            break;
+                        }
+                    }
+                    Ok(ServerMessage::Preempted { .. }) => {
+                        if auto_reconnect {
+                            continue;
+                        }
+                        break;
+                    }
+                    Ok(ServerMessage::Reconnected { .. }) => continue,
+                    _ => continue,
+                }
+            }
+            Ok(Some(Ok(tungstenite::Message::Close(_)))) | Ok(None) => break,
+            Err(_) => break, // timeout
+            _ => continue,
+        }
+    }
+
+    lines
+}
+
+/// Shared async write path used by `AsyncSerialMonitor::write` and
+/// `write_json_rpc`. Sends the already-base64-encoded payload, then
+/// awaits a `write_ack` (with a 5-second bound matching the sync path).
+/// Returns `true` if the ack reported any bytes written.
+pub(crate) async fn write_async(
+    ws_write_slot: Arc<tokio::sync::Mutex<Option<WsSink>>>,
+    ws_read_slot: Arc<tokio::sync::Mutex<Option<WsSource>>>,
+    encoded: String,
+) -> bool {
+    let msg = serde_json::to_string(&ClientMessage::Write { data: encoded }).unwrap();
+
+    {
+        let mut guard = ws_write_slot.lock().await;
+        let Some(sink) = guard.as_mut() else {
+            return false;
+        };
+        if sink.send(tungstenite::Message::Text(msg)).await.is_err() {
+            return false;
+        }
+    }
+
+    // Wait for write_ack. Hold the read half's mutex only across this
+    // single `.next()` so concurrent `read_lines` futures can resume.
+    let mut guard = ws_read_slot.lock().await;
+    let Some(source) = guard.as_mut() else {
+        return false;
+    };
+    let ack_timeout = std::time::Duration::from_secs(5);
+    match tokio::time::timeout(ack_timeout, source.next()).await {
+        Ok(Some(Ok(tungstenite::Message::Text(text)))) => {
+            matches!(
+                serde_json::from_str::<ServerMessage>(&text),
+                Ok(ServerMessage::WriteAck { bytes_written, .. }) if bytes_written > 0
+            )
+        }
+        _ => false,
+    }
+}
+
+/// Async counterpart to `wait_for_remote_json_rpc_response`. Keeps
+/// polling `read_lines_async` until the deadline expires, even if an
+/// individual batch comes back empty — preserving the PR #57 fix.
+pub(crate) async fn wait_for_remote_json_rpc_response_async(
+    timeout_secs: f64,
+    ws_read_slot: Arc<tokio::sync::Mutex<Option<WsSource>>>,
+    auto_reconnect: bool,
+) -> Option<String> {
+    let deadline = std::time::Instant::now() + std::time::Duration::from_secs_f64(timeout_secs);
+
+    while std::time::Instant::now() < deadline {
+        let remaining = (deadline - std::time::Instant::now()).as_secs_f64();
+        if remaining <= 0.0 {
+            break;
+        }
+        let lines = read_lines_async(ws_read_slot.clone(), auto_reconnect, remaining).await;
+        if let Some(json_part) = extract_remote_json_rpc_response(&lines) {
+            return Some(json_part);
+        }
+    }
+
+    None
+}
+
+/// Helper: produce a base64-encoded write payload from raw bytes.
+pub(crate) fn encode_payload(data: &str) -> String {
+    base64::engine::general_purpose::STANDARD.encode(data.as_bytes())
+}
